@@ -6,9 +6,9 @@
     "ng": "ng",
     "build": "npm run env -s && ng build --prod",
     "start": "npm run env -s && ng serve --aot --proxy-config proxy.conf.js",
-    "lint": "ng lint --type-check && stylelint \"src/**/*.scss\" --syntax scss && htmlhint \"src\" --config .htmlhintrc",
+    "lint": "ng lint && stylelint \"src/**/*.scss\" --syntax scss && htmlhint \"src\" --config .htmlhintrc",
     "test": "npm run env -s && ng test",
-    "test:ci": "npm run env -s && npm run lint -s && ng test --single-run --code-coverage",
+    "test:ci": "npm run env -s && npm run lint -s && ng test --code-coverage --watch=false",
     "e2e": "npm run env -s && ng e2e",
     "translations:extract": "ngx-translate-extract --input ./src --output ./src/translations/template.json --format=json --clean -sort --marker extract",
     "docs": "hads ./docs -o",
@@ -22,16 +22,16 @@
     "generate": "ng generate"
   },
   "dependencies": {
-    "@angular/animations": "^5.2.5",
-    "@angular/common": "^5.2.5",
-    "@angular/compiler": "^5.2.5",
-    "@angular/core": "^5.2.5",
-    "@angular/forms": "^5.2.5",
-    "@angular/http": "^5.2.5",
-    "@angular/platform-browser": "^5.2.5",
-    "@angular/platform-browser-dynamic": "^5.2.5",
-    "@angular/router": "^5.2.5",
-    "@ngx-translate/core": "^9.0.1",
+    "@angular/animations": "^6.0.1",
+    "@angular/common": "^6.0.1",
+    "@angular/compiler": "^6.0.1",
+    "@angular/core": "^6.0.1",
+    "@angular/forms": "^6.0.1",
+    "@angular/http": "^6.0.1",
+    "@angular/platform-browser": "^6.0.1",
+    "@angular/platform-browser-dynamic": "^6.0.1",
+    "@angular/router": "^6.0.1",
+    "@ngx-translate/core": "^10.0.1",
     "@ionic-native/core": "^4.5.3",
     "@ionic-native/keyboard": "^4.5.3",
     "@ionic-native/splash-screen": "^4.5.3",
@@ -46,23 +46,25 @@
     "cordova-plugin-whitelist": "^1.3.3",
     "ionic-angular": "^3.9.2",
     "ionicons": "^3.0.0",
-    "core-js": "^2.5.2",
-    "lodash": "^4.17.4",
-    "rxjs": "^5.5.5",
-    "zone.js": "^0.8.18"
+    "core-js": "^2.5.6",
+    "lodash": "^4.17.10",
+    "rxjs": "^6.1.0",
+    "rxjs-compat": "^6.1.0",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
     "cordova": "^8.0.0",
-    "@angular/cli": "~1.7.2",
-    "@angular/compiler-cli": "^5.2.5",
-    "@angular/language-service": "^5.2.5",
-    "@biesbjerg/ngx-translate-extract": "^2.3.1",
+    "@angular/cli": "~6.0.1",
+    "@angular/compiler-cli": "^6.0.1",
+    "@angular/language-service": "^6.0.1",
+    "@angular-devkit/build-angular": "~0.6.0",
+    "@biesbjerg/ngx-translate-extract": "^2.3.4",
     "@ngx-rocket/scripts": "^2.1.0",
-    "@types/jasmine": "^2.5.52",
+    "@types/jasmine": "^2.8.7",
     "@types/jasminewd2": "^2.0.2",
-    "@types/lodash": "^4.14.103",
-    "@types/node": "^8.9.4",
-    "codelyzer": "^4.1.0",
+    "@types/lodash": "^4.14.108",
+    "@types/node": "^8.10.13",
+    "codelyzer": "^4.3.0",
     "hads": "^1.3.4",
     "htmlhint": "^0.9.13",
     "https-proxy-agent": "^2.0.0",
@@ -72,17 +74,17 @@
     "karma-chrome-launcher": "^2.2.0",
     "karma-cli": "~1.0.1",
     "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "^1.1.0",
+    "karma-jasmine": "^1.1.2",
     "karma-junit-reporter": "^1.2.0",
     "protractor": "~5.3.0",
-    "puppeteer": "^1.0.0",
-    "stylelint": "~9.1.1",
-    "stylelint-config-recommended-scss": "~3.1.0",
+    "puppeteer": "^1.4.0",
+    "stylelint": "~9.2.0",
+    "stylelint-config-recommended-scss": "~3.2.0",
     "stylelint-config-standard": "~18.2.0",
-    "stylelint-scss": "~2.5.0",
-    "ts-node": "~5.0.0",
-    "tslint": "~5.9.1",
-    "typescript": "~2.6.2"
+    "stylelint-scss": "~3.1.0",
+    "ts-node": "~6.0.3",
+    "tslint": "~5.10.0",
+    "typescript": "~2.7.2"
   },
   "cordova": {
     "platforms": [
