@@ -2,7 +2,7 @@ import { Component, OnInit, ViewChild } from '@angular/core';
 import { Router, NavigationEnd, ActivatedRoute } from '@angular/router';
 import { Title } from '@angular/platform-browser';
 import { TranslateService } from '@ngx-translate/core';
-import { merge } from 'rxjs/observable/merge';
+import { merge } from 'rxjs';
 import { filter, map, mergeMap } from 'rxjs/operators';
 import { IonicApp, Nav, Platform } from 'ionic-angular';
 import { Keyboard } from '@ionic-native/keyboard';
