@@ -43,10 +43,10 @@ describe('ErrorHandlerInterceptor', () => {
     // Arrange
     // Note: here we spy on private method since target is customization here,
     // but you should replace it by actual behavior in your app
-    spyOn(ErrorHandlerInterceptor.prototype as any, 'errorHandler');
+    spyOn(ErrorHandlerInterceptor.prototype as any, 'errorHandler').and.callThrough();
 
     // Act
-    http.get('/toto').subscribe(() => {}, () => {
+    http.get('/toto').subscribe(() => fail('should error'), () => {
       // Assert
       expect(ErrorHandlerInterceptor.prototype['errorHandler']).toHaveBeenCalled();
     });
