@@ -6,6 +6,7 @@ import { HttpService } from './http.service';
 import { HttpCacheService } from './http-cache.service';
 import { ErrorHandlerInterceptor } from './error-handler.interceptor';
 import { CacheInterceptor } from './cache.interceptor';
+import { ApiPrefixInterceptor } from './api-prefix.interceptor';
 
 describe('HttpService', () => {
   let httpCacheService: HttpCacheService;
@@ -18,6 +19,7 @@ describe('HttpService', () => {
       providers: [
         ErrorHandlerInterceptor,
         CacheInterceptor,
+        ApiPrefixInterceptor,
         HttpCacheService,
         {
           provide: HttpClient,
@@ -41,11 +43,11 @@ describe('HttpService', () => {
   }));
 
   afterEach(() => {
-    httpMock.verify();
     httpCacheService.cleanCache();
+    httpMock.verify();
   });
 
-  it('should use error handler and no cache by default', () => {
+  it('should use error handler, API prefix and no cache by default', () => {
     // Arrange
     let interceptors: HttpInterceptor[];
     const realRequest = http.request;
@@ -60,6 +62,7 @@ describe('HttpService', () => {
     // Assert
     request.subscribe(() => {
       expect(http.request).toHaveBeenCalled();
+      expect(interceptors.some(i => i instanceof ApiPrefixInterceptor)).toBeTruthy();
       expect(interceptors.some(i => i instanceof ErrorHandlerInterceptor)).toBeTruthy();
       expect(interceptors.some(i => i instanceof CacheInterceptor)).toBeFalsy();
     });
@@ -70,7 +73,6 @@ describe('HttpService', () => {
     // Arrange
     let interceptors: HttpInterceptor[];
     const realRequest = http.request;
-    http = http.cache();
     spyOn(HttpService.prototype, 'request').and.callFake(function(this: any) {
       interceptors = this.interceptors;
       return realRequest.apply(this, arguments);
@@ -83,6 +85,7 @@ describe('HttpService', () => {
 
     // Assert
     request.subscribe(() => {
+      expect(interceptors.some(i => i instanceof ApiPrefixInterceptor)).toBeTruthy();
       expect(interceptors.some(i => i instanceof ErrorHandlerInterceptor)).toBeTruthy();
       expect(interceptors.some(i => i instanceof CacheInterceptor)).toBeTruthy();
     });
@@ -93,7 +96,6 @@ describe('HttpService', () => {
     // Arrange
     let interceptors: HttpInterceptor[];
     const realRequest = http.request;
-    http = http.skipErrorHandler();
     spyOn(HttpService.prototype, 'request').and.callFake(function(this: any) {
       interceptors = this.interceptors;
       return realRequest.apply(this, arguments);
@@ -106,7 +108,31 @@ describe('HttpService', () => {
 
     // Assert
     request.subscribe(() => {
+      expect(interceptors.some(i => i instanceof ApiPrefixInterceptor)).toBeTruthy();
       expect(interceptors.some(i => i instanceof ErrorHandlerInterceptor)).toBeFalsy();
+      expect(interceptors.some(i => i instanceof CacheInterceptor)).toBeFalsy();
+    });
+    httpMock.expectOne({}).flush({});
+  });
+
+  it('should not use API prefix', () => {
+    // Arrange
+    let interceptors: HttpInterceptor[];
+    const realRequest = http.request;
+    spyOn(HttpService.prototype, 'request').and.callFake(function(this: any) {
+      interceptors = this.interceptors;
+      return realRequest.apply(this, arguments);
+    });
+
+    // Act
+    const request = http
+      .disableApiPrefix()
+      .get('/toto');
+
+    // Assert
+    request.subscribe(() => {
+      expect(interceptors.some(i => i instanceof ApiPrefixInterceptor)).toBeFalsy();
+      expect(interceptors.some(i => i instanceof ErrorHandlerInterceptor)).toBeTruthy();
       expect(interceptors.some(i => i instanceof CacheInterceptor)).toBeFalsy();
     });
     httpMock.expectOne({}).flush({});
