@@ -7,7 +7,7 @@
 import { enableProdMode } from '@angular/core';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
-import { AppModule } from './app/app.module';
+import { AppModule } from '@app/app.module';
 import { environment } from '@env/environment';
 
 if (environment.production) {
